requirements.txt
flask>=2.0.0
requests>=2.26.0
pillow>=9.0.0
numpy>=1.21.0
discord.py>=2.0.0
asgiref>=3.4.0
python-dotenv>=0.19.0
